Detection of data exfiltration via DNS/requirements.txt
# local package
-e .
collection
math
pathlib 
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
sklearn
pandas
numpy
tldextract
pickle
kafka-python 



